Add special serialize example
