+++ app/examples/custom/serialize/index.jade
@@ -1,0 +1,3 @@
+#vis-element
+#serialize-links
+a#download-link
