+++ app/examples/vis.jade
@@ -1,4 +1,2 @@
 .page
   #vis-container.vis-full
-  #serialize-links
-  a#download-link
